Remove node server in favor of gh-pages web app
After hooking up the finished and working timeclock in the robotics shop,
I discovered it wouldn't work because of the VPN connection to the balena
server required. The school network completely blocks these ports for us
so this commit removes the node server and adds support to pushlish to
github pages instead.

From here, the python code will have to be adjusted to not rely on any
server ports being open, and only communicate with the web app via
the google spreadsheet for purposes of setting an nfc tag.